phase-04/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bleach>=6.3.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.124.4",
    "mcp>=1.0.0",
    "openai-agents>=0.6.4",
    "openai-chatkit>=1.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
